# Conflicts:
#	src/main/java/com/cgvsu/model/Model.java
#	target/classes/com/cgvsu/model/Model.class
#	target/classes/com/cgvsu/objreader/ObjReader.class

+++ src/main/java/com/cgvsu/math/Vector3f.java
@@ -37,7 +37,7 @@ public class Vector3f{
     public float getZ() {
         return z;
     }
-    public double get(int index) {
+    public float get(int index) {
         switch (index){
             case 0: return x;
             case 1: return y;
